requirements.txt
python-dotenv
langchain
langchainhub
langchain_core
langchain_community
langchain_groq
langchain_openai
faiss-cpu
arxiv
wikipedia
streamlit
ddgs